--- CardGameLogic/Enums.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CardGameLogic
-{
-    public enum Suit
-    {
-        Clubs,      //Трефы
-        Diamonds,   //Буби
-        Hearts,     //Червы
-        Spades      //Пики
-    }
-
-    public enum Player
-    {
-        Bot,
-        Human
-    }
-}
